+++ src/main/java/jp/main/taikun/mysticaleverything/config/IFilter.java
@@ -1,0 +1,7 @@
+package jp.main.taikun.mysticaleverything.config;
+
+import net.minecraft.world.item.ItemStack;
+
+public interface IFilter {
+    public boolean filter(ItemStack itemStack);
+}